+++ src/three/gltf/classes/ClassProperty.js
@@ -1,8 +1,8 @@
-import { getField, resolveDefault } from './PropertySetAccessor.js';
+import { adjustValue, getField, resolveDefault, resolveNoData } from './Utils.js';
 
 export class ClassProperty {
 
-	constructor( property ) {
+	constructor( enums, property, accessorProperty = null ) {
 
 		this.type = property.type;
 		this.componentType = property.componentType || null;
@@ -18,12 +18,45 @@ export class ClassProperty {
 		this.noData = getField( property, 'noData', null );
 		this.default = getField( property, 'default', null );
 		this.semantic = getField( property, 'semantic', null );
+		this.enumSet = null;
+
+		if ( accessorProperty ) {
+
+			this.offset = getField( accessorProperty, 'offset', this.offset );
+			this.scale = getField( accessorProperty, 'scale', this.scale );
+			this.max = getField( accessorProperty, 'max', this.max );
+			this.min = getField( accessorProperty, 'min', this.min );
+
+		}
+
+		if ( property.type === 'ENUM' ) {
+
+			this.enumSet = enums[ this.enumType ];
+			if ( this.componentType === null ) {
+
+				this.componentType = getField( this.enumSet, 'valueType', 'UINT16' );
+
+			}
+
+		}
 
 	}
 
 	resolveDefault( target ) {
 
 		return resolveDefault( this, target );
+
+	}
+
+	resolveNoData( target ) {
+
+		return resolveNoData( this, target );
+
+	}
+
+	adjustValueScales( target ) {
+
+		return adjustValue( this.type, this.componentType, this.scale, this.offset, this.normalized, target );
 
 	}
